refactor(workflow): split workflow.py into package + add transition_id_unique helper
Norman flagged workflow.py (521 lines) as exceeding the 500-line readability
budget. Steve flagged a missing transition_id uniqueness helper symmetric to
_validate_stage_id_unique. Both fixes land together since they touch the
same file boundaries.

Package split (each file under 500 lines, max 272):

  backend/src/bakufu/domain/workflow/
    __init__.py         (60 lines) — public re-exports incl. _validate_*
    dag_validators.py   (272)      — 10 module-level invariant helpers
    entities.py         (120)      — Stage / Transition Pydantic models
    workflow.py         (218)      — Workflow Aggregate Root only

The directory split is the *physical* ground for Confirmation F's twin-
defense: aggregate-level helpers cannot share code with Stage self-validation
because they don't even share a file. __init__ re-exports the underscore-
prefixed helpers so tests' direct-import contract (TC-UT-WF-060) is preserved
without changing test code.

New helper _validate_transition_id_unique:

* Symmetric to _validate_stage_id_unique
* New kind 'transition_id_duplicate' added to WorkflowViolationKind Literal
  with docstring noting the symmetry rationale
* Dispatched second-after stage uniqueness in Workflow._check_invariants so
  duplicate-id detection happens before any DAG traversal
* MSG follows the Stage convention: '[FAIL] Transition id duplicate: <id>'

160 existing tests still pass; new helper added test by ジェフ in follow-up.
Coverage 98% overall, ruff clean, pyright strict 0 errors.

Refs #9
